LPS-54098 Rename package
