+++ portal-impl/test/integration/com/liferay/portal/service/OrganizationLocalServiceTreeTest.java
@@ -21,7 +21,7 @@ import com.liferay.portal.kernel.test.util.TestPropsValues;
 import com.liferay.portal.model.Organization;
 import com.liferay.portal.model.OrganizationConstants;
 import com.liferay.portal.model.TreeModel;
-import com.liferay.portal.service.test.BaseLocalServiceTreeTestCase;
+import com.liferay.portal.service.tree.test.BaseLocalServiceTreeTestCase;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.MainServletTestRule;
